Add expect/actual for YoutubeExtractor
NewPipeExtractor won't work in non-JVM platforms. There is a KMP version of it but it is to small and I don't think we should really use it right now.
